requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dictdiffer==0.9.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pydantic==1.10.11
PyJWT==2.7.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
